Fix lint issue - separate test fragment
