+++ Corona-Warn-App/src/deviceForTesters/java/de/rki/coronawarnapp/test/contactdiary/ui/ContactDiaryCommentInfoTestFragment.kt
@@ -1,14 +1,15 @@
-package de.rki.coronawarnapp.contactdiary.ui.day.tabs
+package de.rki.coronawarnapp.test.contactdiary.ui
 
 import android.os.Bundle
 import android.view.View
 import androidx.fragment.app.Fragment
 import de.rki.coronawarnapp.R
 import de.rki.coronawarnapp.databinding.ContactDiaryCommentInfoFragmentBinding
+import de.rki.coronawarnapp.test.menu.ui.TestMenuItem
 import de.rki.coronawarnapp.util.ui.popBackStack
 import de.rki.coronawarnapp.util.ui.viewBindingLazy
 
-class ContactDiaryCommentInfoFragment : Fragment(R.layout.contact_diary_comment_info_fragment) {
+class ContactDiaryCommentInfoTestFragment : Fragment(R.layout.contact_diary_comment_info_fragment) {
 
     private val binding: ContactDiaryCommentInfoFragmentBinding by viewBindingLazy()
 
@@ -18,5 +19,13 @@ class ContactDiaryCommentInfoFragment : Fragment(R.layout.contact_diary_comment_
         binding.toolbar.setNavigationOnClickListener {
             popBackStack()
         }
+    }
+
+    companion object {
+        val MENU_ITEM = TestMenuItem(
+            title = "Contact Diary Comment Info",
+            description = "Contact diary comment info screen",
+            targetId = R.id.test_contact_diary_comment_fragment
+        )
     }
 }
